Fix init
Prior to this, every time this config would be loaded on a new machine,
the user would have a flood of errors about this that and the other
thing.

Now, we should only have to clone the repo, run nvim once, and then
reopen it after all the treesitter compilations are done.

This seems to work consistently.